wxGUI: fix for wxPython 4, see https://trac.osgeo.org/grass/ticket/3621
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73404 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ gui/wxpython/psmap/utils.py
@@ -175,9 +175,9 @@ def convertRGB(rgb):
         return str(rgb.Red()) + ':' + str(rgb.Green()) + ':' + str(rgb.Blue())
     # transform a GRASS named color or an r:g:b string into a wx.Colour tuple
     else:
-        color = (grass.parse_color(rgb)[0] * 255,
-                 grass.parse_color(rgb)[1] * 255,
-                 grass.parse_color(rgb)[2] * 255)
+        color = (int(grass.parse_color(rgb)[0] * 255),
+                 int(grass.parse_color(rgb)[1] * 255),
+                 int(grass.parse_color(rgb)[2] * 255))
         color = wx.Colour(*color)
         if color.IsOk():
             return color
